requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-core==2.23.4
pytest==8.3.3
pytest-alembic==0.11.1
python-dotenv==1.0.1
pywin32==307
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tomli==2.0.2
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
